+++ src/style/index.ts
@@ -1,0 +1,2 @@
+export { initFadeLeftMotion, initFadeMotion, fadeIn, fadeInLeft, fadeOut } from './motion/fade';
+export { initMotion } from './motion/init';